+++ avatars/client.py
@@ -1,5 +1,5 @@
 # This file has been generated - DO NOT MODIFY
-# API Version : 0.4.4
+# API Version : 0.4.5
 
 import sys
 from collections.abc import Mapping, Sequence